First commit.
This is the first commit. Broke out the classes into their own files.
