+++ PhysiPy/__init__.py
@@ -1,0 +1,20 @@
+__all__ = [
+    "constants",
+    "derivations",
+    "Electricity",
+    "Electromagnetism",
+    "Electrostatics",
+    "Errors",
+    "equations",
+    "FluidStatePhysics",
+    "Gravitation",
+    "Mechanics",
+    "Nlm",
+    "QuantumMechanics",
+    "SolidStatePhysics",
+    "Subatomic",
+    "Thermodynaics",
+    "Waves"
+]
+
+
