+++ requirements.txt
@@ -1,0 +1,2 @@
+numpy~=1.26.2
+setuptools~=57.0.0
